Extension of the classical PyTorch Model
Added a quantum layer to the output of "fc3" layer of the original classical model (see the Conspicuity_Detection.ipynb notebook) to see how an extra quantum layer affects the model and trained it. I am unsure if it is properly working, even though it is giving some crazy results.